+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Optimize:optimize SpringValueProcessor.](https://github.com/Tencent/spring-cloud-tencent/pull/655)
+- [Optimize/optimize annotation.](https://github.com/Tencent/spring-cloud-tencent/pull/672)